[iOS] Cancel banner requests when modals are completed.
This CL creates InfobarModalCompletionNotifier, which notifies observers
when all the OverlayRequests for modal UI for an InfoBarIOS are
completed.  This is owned by InfobarOverlayRequestInserter, and is used
by InfobarBannerOverlayRequestCancelHandler to cancel banners after
the completion of the detail sheet and modal UI for that InfoBarIOS is
finished.  To facilitate this animation, the default behavior for UI
cancellation via OverlayPresenter was updated to become animated.

This CL also creates FakeOverlayRequestCancelHandler, which is used
by OverlayRequestQueueImplTest and InfobarModalCompletionNotifierTest.

Bug: 1030357
Change-Id: If0bb73537ea6d7304725a580b0de27b345dbb8de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2031767
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#740185}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e973a0111d2293a856e2c1f50e7e3686f1641601