[iOS] Cleanup names for TabGrid animations
This CL updates the names of the object providing animations
information from GridTransitionStateProviding to
GridTransitionAnimationLayoutProviding and its methods.
Hopefully, it should increase the readability of the responsibilities
of the object.

It also removes the context parameter as it wasn't used and I plan to
use this class without a context in a future CL.

Bug: 1038034
Change-Id: I9b125cdd6603c7441dfd0cb9f1c256b3d60e37b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2050317
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#740327}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 273c468d7e7dafa3361ea5d56f2bbe7c0e102372