[iOS] Fail gracefully for unsupported infobar overlay types.
The previous implementation of InfobarOverlayRequestInserter and
InfobarOverlayRequestFactoryImpl crashed if an unsupported infobar
overlay was requested.  This ultimately should be the case, as it
helps future users of these classes to remember to add the request
factory.

However, when there are infobars that haven't been converted yet, this
requirement makes debugging difficult becuase the app crashes when the
session restoration infobar is added.  This CL replaces this crash
behavior with early returns to make debugging easier before the remaining
infobars are converted.

Bug: 1030357
Change-Id: I68ef5b758444e65b0985d970c92de0bf935785cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2041032
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#741296}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f405f017b1757bb785ff1760fdcbc94b5af2f90d