[iOS] Add confirmation and interrupt buttons to user sign-in migration.
This patch is a follow-up from crbug.com/2041599 that adds the button
to continue the sign-in operation and the button to cancel the flow.

This CL is part of a series of refactors to move the existing sign-in
architecture to the Coordinator-Mediator design paradigm.

Fore more information, see go/chrome-ios-signin-migration.

Bug: 971989
Change-Id: Icb2f77f7a4016cb008d4b808af7784e45dd41f18
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2054208
Commit-Queue: Nohemi Fernandez <fernandex@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#741868}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: af5590597bd3f21415a98998825a707d0aa309ed