[iOS] Fix initialization bug in ActiveWebStateObservationForwarder
The bug occurs when an instance of ActiveWebStateObservationForwarder
is created while already in a WebStateActivatedAt callback. In this case,
the new instance is already observing the newly active webstate, but will
receive its own WebStateActivatedAt callback. Because it is already
observing the new webstate, it doesn't need to stop or start observation.

Bug: none
Change-Id: I80826302e17c828874d49bc9cd6957c29eeb6f08
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2059371
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#741944}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6fb456dc246a4d76ed67fa1ae06d0ba091795d6e