[iOS] Remove launch mask when the BVC is contained
The launch mask was used to prevent the user from seeing the TabGrid when
a tab should have been presented as it is not possible to directly show
the BVC when it is presented.
Having it contained allow to remove this workaround.

Bug: 1038034
Change-Id: I2cc74afe5d00406d1be4de20ed7e96f50cd839a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2061759
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#742141}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6c18c060bd76a220433af61116d84163c5bb0581