[iOS] Clean up BVC's |-setUpViewLayout:|.
This CL combines several block for the initial layout pass into a single
block.  Additionally, this CL updates the layout of the primary toolbar
layout so that it is behind the tab strip on iPad and in back of the
typing shield on iPhone.  This more explicitly describes the semantic
purpose of the previous |bottomView| logic.

Bug: none
Change-Id: I272882c4958afd2cdc25694a508bfe9119cb7db4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2045212
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Auto-Submit: Kurt Horimoto <kkhorimoto@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#742322}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ecc9f592260082c09a201cae7bf62d940f6f4864