[iOS] Adding DCHECK for -SigninCoordinator.signinCompletion
|signinCompletion| has to be set by the owner of the SigninCoordinator
to track when the the sign-in is done, when the coordinator can be
released.
Adding DCHECK(self.signinCompletion) in -[SigninCoordinator start], to
enforce it.

Once the completion block is called, the property is set to nil.
Adding DCHECK(!self.signinCompletion) in -[SigninCoordinator stop], to
make sure the callback is called before stopping the coordinator.

Bug: 971989
Change-Id: I0bfcc2579a8824bc6e4b8345221542bcf5819359
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2053129
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#742331}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 272b75fe04d4d86672a729b16ba943db3c8217b7