[iOS] Disable voice search while VoiceOver is enabled.
When VoiceOver is enabled, triggering the voice search UI begins reading
the accessibility elements on the screen.  There is not a good way to
check when VoiceOver has stopped speaking, so voice search cannot simply
wait for VoiceOver to stop before recording.  Moreover, removing the
accessibility features for the other views in the voice search UI is
also not a solution, as VoiceOver users would be stuck on the voice
search screen.  The long-term solution to this issue is to update the
voice search feature to ignore audio that is emitted from the device,
but this fix is not yet built.

Instead, this CL disables voice search while VoiceOver is enabled.
Visually-impaired users can still use other iOS accessibility features
to dictate queries into the omnibox, so this does not negatively
affect the ability of these users to search.

VoiceSearchAvailability was created to implement this feature.  It is
an observable helper object that uses VoiceOver notifications and
VoiceSearchProvider::IsVoiceSearchEnabled() to determine whether voice
search should be enabled.  The NTP and keyboard accessory view were
updated to use VoiceSearchAvailability to optionally disable the
microphone button.

Bug: 998524
Change-Id: I8789a9a05542bdc659f8d98d0f1c90de3dc3ecb1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2005903
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Auto-Submit: Kurt Horimoto <kkhorimoto@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#742372}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fd40aa3d0cf1879ad6e683e116361336c9541599