[ios] Handle intents triggered in the foreground too.
The intent handler for SearchInChromeIntent and OpenInChromeIntent both
assumed the intents would only happen while the app was backgrounded.
This worked since that's usually true, and the startupInformation would
be processed when the app was foregrounded.

However, OpenInChromeIntents can be chained, and presumably this is also
true for SearchInChromeIntent, which means the first in the chain will
trigger while the app is backgrounded, and the remainder when the app is
in the foreground.

Instead of assuming the foregrounding step will process
startupInformation, let the logic flow to -continueUserActivityURL.

Bug: 1051783
Change-Id: Ia61334ac33995030744c1316f5accef2eae15b21
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2062097
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Justin Cohen <justincohen@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#743160}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e5e97fd74960858a54206c273e9a7985e9ee352e