[ios] Controls iOS enterprise policy via a command line flag.
The policy system is initialized before about:flags or field trials, so
it is not possible for either of those mechanisms to decide whether or
not policy should be enabled at runtime. Instead, we will switch to
looking for the presence (or absence) of a command line flag. This will
allow us to easily turn policy on locally during development, and when
we are ready to ship policy support, we can remove the command line
flag.

This partially reverts commit 8eefc0a126a10305d58cf7acdbae6af4313eb3b2.

BUG=1024115

Change-Id: I0839aa1ab0bd68b149d900dfe522b6d4081b5e49
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2068681
Commit-Queue: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#743847}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bbae1e0f9c6ce878ecd8b21e04eb1a3f5f14235a