[iOS] Refactor tab restoration for multi-window.
This CL refactors the tab restoration classes for multi-window, scoping
some functions to a Browser, and removing use of TabModelList.

Non-functional changes:

* TabRestoreDelegateImplIOS (which is the iOS implementation of the
  sessions::LiveTabContext API, not "TabRestoreDelegate") was a keyed
  service. It's renamed to LiveTabContextBrowserAgent and (accordingly)
  scoped to a Browser (on desktop, Browser directly implements this
  API.
* The associated factory class is removed.
* TabRestoreServiceClient is updated to use Browser and BrowserList
  instead of TabModel and TabModelList.
* The RestoreTab() helper function is updated to take a Browser param
  instead of a BrowserState.

FUNCTIONAL CHANGE:

Previous to this CL, IOSChromeTabRestoreServiceClient's implementation
of CreateLiveTabContext() simply returned the (single) BrowserState-
scoped sessions::LiveTabContext implementation. Looking at the desktop
implementation of this function, it appears to create a new window:

https://cs.chromium.org/chromium/src/chrome/browser/sessions/chrome_tab_restore_service_client.cc?type=cs&g=0&l=35

While that might be a desirable behavior when multi-window is supported,
we don't yet have a set of hooks available to do that cleanly; nor
do we want to instead return some arbitrary, already-opened window.

It looks like the call sites for this (in TabRestoreServiceHelper) are:

 - When restoring a session by ID with the WINDOW type, and
 - When restoring a tab without a provided context or with the NEW_WINDOW
   disposition.

ios/ doesn't currently have code paths that do either of these, and this
CL adds a check in the RestoreTab helper to guard against them. Then, the
implementation of CreateLiveTabContext() now is a NOTREACHED() no-op.

crbug.com/1056596 tracks adding support for window restoration with
multi-window.

Bug: 1056208
Change-Id: Ib51ea7ea6efd3a51e6ec07d031e1a3ae1203579b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2078915
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Mohammad Refaat <mrefaat@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#746384}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 77bef0e74e398ba8f8bfc11ad5a6b28a9c412928