[iOS] Remove TabModelList from AutocompleteProviderClientImpl
This straightforwardly replaces the use of TabModelList to use
BrowserList. Technically this is a functional change because it will
check if the passed URL is open in any tab in any window, but that was
presumably the desired behavior in the first place.

Bug: 1058332
Change-Id: I52075acd0b6e3fa0ccd3bad69ee810769f1031f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2087333
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Mark Cogan <marq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#747185}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: aea1645c1b0644aed7c281849d21182d5bd4c692