[ios] Create UKM pageload duration TabHelper
This TabHelper will observe WebStateObserver and record page sessions
while its WebState is the active WebState (i.e. it is visible on the
screen). When a new page is navigated to, the app backgrounds, or if
the WebState is destroyed, the recording will stop and a UKM will be
logged for the page session duration.A new IOS.PageNavigation UKM will
be logged for every navigation completion.

Bug: 1006357
Change-Id: I9c7184a1cc3d8f9ff1a39f34f3363a388738e82a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2068244
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Reviewed-by: Annie Sullivan <sullivan@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#747328}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 05e6d5e5f346dadd501bac4a64042caf5c31f626