[iOS] Fix BVC contained animation on first run
The first run animation is done with the BVC being shown and the
FirstRunViewController being presented on the same runloop.
On official, this calls the completion handler of the BVC animation
with finished at NO even if it actually completes.

For now this CL fixes it by ignoring the finished value as it is not
taken into account anywhere else (so it is not working).
The long term solution will be to have a better way to present the
first run.

Fixed: 1056697
Change-Id: Iaf30240f4da98cf9d5856d34b6be174b371f1ccd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2089767
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#748138}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6dbe703aa21e05b6e2bceeb4f5314a0bf2e1011e