[iOS] Fix animation glitch on TabGrid
This CL removes the animation of the content offset of the TabGrid which
allows the selected tab to be at the top of the collection.
When BVC is presented the animation isn't seen. When it is contained,
the animation was creating a visible glitch.

Fixed: 1056611
Change-Id: I8b32e8191c413e368f75c88dce1fee0665f6d075
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2095722
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#748659}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 18e82a3bfb2fee796c6241db5df3f8b41d45d53c