[iOS] Migrate away from downstream URLLoadingService use (3/8)
As part of preparing for multi-window, the URLLoadingService needs
to be migrated to be a BrowserAgent instead of a KeyedService. That
change is itself fairly large, but as an additional complication
URLLoadingService is used in downstream code.

This CL is the third in a series of changes needed to complete this
refactor:

1. (crrev.com/c/2089687) Add a new provider API upstream that passes
   both a Browser and temporary URL loading shim to CreateLogoVendor.
2. Implement this new API downstream and add support for initializing
   the LogoController with both Browser and the loading shim. Use the
   loading shim for loading URLs if it is present.
3. Upstream, call the new CreateLogoVendor() API.
4. Downstream, remove support for the old CreateLogoVendor() API and
   update LogoController to only be initialized with a Browser and the
   shim loader. Remove downstream use of the URLLoadingService.
5. Upstream, refactor URLLoadingService to be URLLoadingBrowserAgent,
   including updating the upstream implementation of the loading shim
   passed downstream in CreateLogoVendor().
6. Upstream, add a new provider API that no longer passes the loading
   shim into CreateLogoVendor().
7. Downstream, remove use of the loading shim from LogoController and
   have it fetch the URLLoadingBrowserAgent from its Browser parameter.
   Implement the new shim-less CreateLogoVendor() API. Remove the
   implementation of the shimmed API.
8. Upstream, remove the loading shim protocol and the old shimmed
   CreateLogoVendor() API. Call only the new shimless API.

Specifically this CL changes the contents suggestion coordinator to
call the CreateLogoVendor API added in CL #1, and to implement the
shim URL loading protocol. It also adds the upstream provider API
implementation that should have been in place in CL #1, and hardens
ContentSuggestionsHeaderViewController against a nil logo vendor.

(Also, as a minor note, this CL removes the synthesis of -setLogoVendor:
in ContentSuggestionsHeaderViewController, which was confusing and
a mismatch for the protocol definition which doesn't define it as a
property).

(This CL should not land until downstream CL #2 has landed and rolled).

correctly

TEST: Manually set #use-ddjson-api and tested that doodle links open
Bug: 1046374
Change-Id: Ibf971db36daf9ba35196cd424f28e637b8d2663b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2089785
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#748671}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9f0aec1999a0ef056a62e12a8a34fc44d751e311