[ios] Adds HistoryTVC empty tableFooterView.
This was removed in https://crrev.com/c/2058083, but for History's case
this is still needed in order to get rid of the separators when the
tableview is empty.

Other collections hide the separator lines when the screen is shown,
since History has a search function is a bit different and this was
the easiest safest way of doing it.

Bug: 1057998
Change-Id: Ia58ee0bc70ad08daf19c1d0721d894986de4e333
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2090915
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#748792}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4b537624a82fc943482528ff552f724a6a80edf5