Make the coverage config reusable for generic profiling.
Most of the coverage configuration can be re-used for general
Clang-based profiling (e.g. the profiling step of a PGO build).
This changes most of the |use_clang_coverage| checks to
|use_clang_profiling|.

The toolchain opt-out option is still limited to the coverage
builds (grep for "toolchain_use_clang_coverage" in CodeSearch).
I.e. targets built simply with |use_clang_profiling=true| won't
be opted out by the clang_code_coverage_wrapper.py script.

This CL is a pure refactoring with no functional change.
Getting profiling information from the sandboxed renderers work
this way:
- After creating a new child process the browser will open a
  profiling file for this process and pass the handle to the
  child via the SetProfilingProfile Mojo call (see change in
  content/browser/renderer_host/render_process_host_impl.cc).
- This process will used by the LLVM instrumentation to record
  profiling information.
- When terminating, child processes will call
  WriteClangProfilingProfile to write the data to the profiling
  data file.
- Once all the raw profile data files have been collected,
  they'll be merged (via the "llvm-profdata" tool) and used.

Note that this isn't meant to be enabled into any build that
will be shipped to users (a profiling build is usually 10x
bigger than a normal build anyway, not really shippable).

The goal of this refactoring is to make it possible to reuse
this mechanism for the instrumentation step of PGO builds.

Bug: 1058979
Change-Id: If8321234f6b6f86befd81f6e9e068a342728d605
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2090493
Reviewed-by: Will Harris <wfh@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Yuke Liao <liaoyuke@chromium.org>
Reviewed-by: Sajjad Mirza <sajjadm@chromium.org>
Commit-Queue: Sébastien Marchand <sebmarchand@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#749015}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c7695f63126642368835ec006f9e53fe913607f0