[iOS] Create OverlayPresentationContextCoordinator
This coordinator presents a UIViewController over the context of the
OverlayContainerViewController.  This is necessary because custom
UIViewController presentation with a UIPresentationController whose
|shouldPresentInFullscreen| is NO results in the presented container
view to be added to the presentation container view of the nearest
presented UIViewController.

Although OverlayContainerViewController's |definesPresentationContext|
returns YES, the container view is added as a child UIViewController
in the view hierarchy.  This means that custom UIViewController
presentation cannot use its presentation context.  To fix this,
OverlayPresentationContextViewController is presented using
UIModalPresentationOverCurrentContext so that a UIKit-provided
presentation context is inserted into the view hierarchy over the
container view's modality.

This CL updates the OverlayPresentationContextImpl to be based on
the container and presentation context view controllers to update
its presentation capabilities so that overlay UI presented with
UIModalPresentationCustom can be correctly scoped to the appropriate
context.

Since the UIKit-provided presentation context intercepts touches by
default, the OverlayPresentationContextViewController's view is
resized to CGRectZero when there is no visible overlay UI.  This
allows touches that hit the container view when there are no overlays
displayed to be forwarded to the underlying browser UI.  The hit test
and touch forwarding code used in OverlayContainerView cannot be used
here since the touch handling for the UIKit-provided presentation
container view cannot be customized.

This CL also creates OverlayPresentationController, which can be used
to communicate presented UIViewController presentation information to
the presentation context.  This is used to resize the presentation
context's container view such that it exactly matches the presented
overlay UI.  This will be used by infobar banner overlays so that
touches that lie outside of the banner can pass through to the browser
UI.

Additionally, this CL removes the DCHECK for BVC's snapshot overlay
calculation since it is no longer appropriate.  Previously, overlays
could only be displayed as a child of the container or as presented
upon it.  After this CL, the container has the presentation context
permanently presented upon it and this presented view controller is
used as the base for any overlay UI implemented using presented
UIViewControllers.

Bug: 1030357, 1054344
Change-Id: I170add2bfc1bc37a9c211f422ab4c494f47b67ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2038252
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#749232}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 94e76188bcac0d07c628f5fa745a973f2e1b137f