[iOS] Make a Safe Browsing query on each navigation
This CL introduces a SafeBrowsingTabHelper, which is a WebStatePolicyDecider
that makes a Safe Browsing query on each call to ShouldAllowRequest(). For
now, the result of the Safe Browsing query is ignored; the purpose of making
the Safe Browsing query is to exercise the Safe Browsing database code and
record UMA (being added in http://crrev.com/c/2087415) for database freshness.

This new behavior is behind the kSafeBrowsingAvailableOnIOS flag.

Bug: 1049340
Change-Id: I2844637fa9e3c0845fca6355ff1547b9d19252c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2092127
Commit-Queue: Ali Juma <ajuma@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Varun Khaneja <vakh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#749343}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 65d98ed3c670f93d919342c3c64d46d4a61928d3