[Autofill Auth] Clank checkbox is shown only if GDFGRP allows opt-in.
Currently the checkbox on the CVC prompt is always shown if the flag is turned
on. However, it should be hidden in the case that the user is not verifiable
or GetDetailsForGetRealPan returns false for `offer_fido_opt_in`.

The information of whether or not to show the checkbox on the card unmask
prompt on Android is stored in CreditCardAccessManager. This information needs
to be propagated to CardUnmaskPrompt. The path is as follows --

1. CreditCardAccessManager (implements CreditCardCVCAuthenticator::Requester)
2. CreditCardCVCAuthenticator (implements FullCardRequest::UIDelegate)
3. FullCardRequest (implements CardUnmaskDelegate)
4. CardUnmaskPromptController
5. CardUnmaskPromptViewAndroid
6. CardUnmaskBridge (Java)
7. CardUnmaskPrompt (Java)

Bug: 949269
Change-Id: I5b747856604750f8a30da066ffda84e9c3cda7ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2087983
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Reviewed-by: Jared Saul <jsaul@google.com>
Commit-Queue: Manas Verma <manasverma@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#749925}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 122de32bc9042bd31bf142dee6cd69a0154a27a6