[iOS] Implement custom presentation for infobar modals.
This CL updates InfobarModalOverlayCoordinator to use an
InfobarModalTransitionDriver to present infobar modal UI.

It also updates the save password infobar delegate so that
calls to InfobarPresenting() and InfobarDismissed() are
idempotent.  This is necessary because with the second
phase of the infobar UI reboot, the modal is presented
while the banner is still visible if the banner's show
modal button is tapped.  This should only count as a single
infobar presentation event.

Bug: 1030357
Change-Id: I3266395c25215163d32bf4d669e906804df172e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2041031
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#750378}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 111d22eca5e0abf0c2c26c13c1f70c6633168519