[iOS] Add utility class for presenting sign-in promo.
As a part of the sign-in architecture migration we are moving logic
for presenting the sign-in promo out of a custom view controller and
into a utility class.

Logging from this custom view controller will be added subsequently
to a specialized UpgradeSigninLogger class.

Bug: 971989
Change-Id: I83aeb58ccaaff75d833015597e3ec2cb8992721d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2107552
Commit-Queue: Nohemi Fernandez <fernandex@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#752495}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c80e46d29bbd2655a049180c34785d21ae66b89b