[iOS] Convert DeviceSharingManager to a KeyedService
Updates the device sharing manager to be a keyed service.

The device sharing manager implementation was such that it would be
more cleanly implemented as a keyed service; prior to this CL the
service needed to be updated manually when browser states were created
or destroyed.

By making this change, explicit handling of device sharing manager
creation, clearing, and destruction is removed, simplifiying the
browser wrangler and startup/shutdown APIs.

This change re-implements the Objective-C device sharing manager as a
C++ keyed service class with an accompanying factory class. The basic
behavior of the class remains unchanged, with the following exceptions:

  - There is no longer an "update for browser state" or "clear device
    sharing manager" method. As a keyed service, the manager is bound
    to a specific browser state for its lifetime.
  - Instead of expecting calling code to set an empty URL to indicate
    that no URL is being shared, an explicit ClearActiveUrl() method is
    added.

The device sharing manager implementation (as before) creates and owns
an instance of HandoffManager, which in turn creates user activities
that expose the active URL. For testing purposes, the handoff manager
and its activity's URL are exposed; the prior implementation mocked
this instead. For EGTests, the utility that grabbed the device sharing
manager from MainController is removed, and instead the app interface
uses the keyed service to access it (also using the test-exposed
interface for the activity URL).

This CL overall is the first step in a two-part process to update
device sharing for multi-window; the second CL will factor the sharing
logic out of BrowserViewWrangler into a BrowserAgent.

Bug: 1049535
Change-Id: If067547fcaf7caabb334adf25c7681fa286c53f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2115610
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Mark Cogan <marq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#752501}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ec3c59f359e398bd1f9ce7b617717bed02bfa2a7