[iOS] Create DeviceSharingBrowserAgent
This CL creates a browser agent to manage device sharing, and updates the
device sharing system to consistently support multi-window.

There are two fundamental changes involved.

First, the device sharing service now has the notion of an "active"
browser. All calls to update or clear the active URL for the service
are issued in the context of a browser, and if that browser isn't the
active one, no change is made to the shared URL.

Second, all of the logic in BrowserViewWrangler to track which URL is
active is moved to the new browser agent. The active URL is updated when
new web states are activated in the browser, or when the active web state
completes a navigation. BrowserAgents for incognito browsers always
clear the active URL when it would be updated.

Every browser has a device sharing browser agent created; the device
sharing manager then regulates which updates are propagated to the
single handoff manager for the browser state, based on preference
settings and which browser is active.

BrowserViewWrangler's only responsibility for device sharing is now to,
when an interface is marked as the current interface, mark that as the
current interface. Once a more robust scene lifecycle is available in
SceneController, that responsibility can move to the appropriate event
hook. (Currently, just switching focus between scenes will not update
the shared URL).

Change-Id: I117778789aea88c6625f06576f6fa04cbdb38b3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2115479
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#752820}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b70e2e42878a0b99bebe691ad7f29f921dcc3027